Majority Of Settings Stuff Implemented, 8/24/2024
Majority Of Settings Stuff Implemented. So got quite alot of work done on this mod today. Implemented most of the remaining stuff I had planned to put into it for now, as well as getting most of the settings stuff implemented and logic done for all of that. Main thing I need to do now is test, rewrite settings description text, and organize settings and code and such, after that this should basically be ready for v1.00 release, 8/24/2024.